requirements.txt
Flask
Flask-WTF
Flask-JSON
flask-sqlalchemy
flask-bcrypt
flask-login
gunicorn